# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Modern Todo Manager API",
    "version": "1.0.0",
    "description": "REST API for managing Todos, powered by Next.js and Convex."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "schemas": {
      "Todo": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": ["Low", "Medium", "High"]
          },
          "dueDate": {
            "type": "number",
            "description": "Unix timestamp"
          },
          "completed": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "number"
          },
          "updatedAt": {
            "type": "number"
          }
        }
      },
      "CreateTodoInput": {
        "type": "object",
        "required": ["title", "priority"],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": ["Low", "Medium", "High"],
            "default": "Medium"
          },
          "dueDate": {
            "type": "number"
          }
        }
      },
      "UpdateTodoInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": ["Low", "Medium", "High"]
          },
          "dueDate": {
            "type": "number"
          },
          "completed": {
            "type": "boolean"
          }
        }
      }
    }
  },
  "paths": {
    "/api/todos": {
      "get": {
        "summary": "Get all Todos",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["All", "Active", "Completed"]
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["Newest First", "Oldest First"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of todos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new Todo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/todos/{id}": {
      "get": {
        "summary": "Get a Todo by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Todo object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a Todo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a Todo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted successfully"
          }
        }
      }
    }
  }
}